feat(cssparts): add support to customize content area #46
Multiple PRs have been created (#33,#40) to help support custom content.
The issue remains that this issue is too difficult to manage this way
and a third issue was created to support the exact same scenario.

As the maintainer of a design system, trying to support all styles of
content is not possible. We can do our best to support use cases
that we are aware of, but in the case where the content does not
fit the model we support, we must allow our users a way to customize
per their content, and that is the role of CSS Parts.

Changes to be committed:
modified:   demo/index.html
modified:   docs/api.md
modified:   src/auro-accordion.js
modified:   src/style.scss