fix(trigger content): update content alignment in trigger #47
The scope of this commit is to update the treatment of content
in the trigger slot of the accordion. This solution will provide
an initial opinion of alignment while allowing users to customize
the space per their specific content needs.

Changes to be committed:
modified:   demo/index.html
modified:   src/auro-accordion.js
modified:   src/style.scss