docs: remove custom component content from demo page
This commit removed the custom extension of a component from the demo
page as this is currently not working and is displaying a broken UI to
the users of the element.

This commit will be referenced in a new issue to address this
unformatted UI issue.

Changes to be committed:
modified:   demo/demo.md
modified:   docs/partials/demo.md