docs: update install directions
The partial was updated to better render the code example as well remove
the reference to the ES5 version of the bundled module as it is NOT
necessary to be installed any longer due to IE11 no longer being a
supported browser.

The content was also updated to reflect the transition from the v1.6.6
version to the new v3.0 version.

Changes to be committed:
modified:   README.md
modified:   docs/partials/readmeAddlInfo.md